Removed player build handler from AddressablesEditor and refactored actions to be performed in a preprocessor class. Also changed it to detect both batch or headless builds.
